Refactor to use GenServer instead of Agent
Needing to use GenServer so that we're doing the right thing w.r.t.
async tasks. Specifically, GenServers make it easy to define a handler
for a generic message, and we can use the Process.send_after command to
trigger game phase transitions (rather than an async Task process with
a sleep on it). This is a bit easier to understand and seems more
correct. Additionally, the use of a GenServer more cleanly separates the
state alteration code from the public interface, in my opinion. (Agent
seems more built to serve a simpler use case with simpler state
transitions).